requirements.txt
Flask
psycopg2
gunicorn
Flask-Bcrypt
Flask-Cors
PyJWT